+++ crates/uv-resolver/src/candidate_selector.rs
@@ -1,7 +1,3 @@
-use std::cmp::Reverse;
-use std::collections::BinaryHeap;
-use std::iter::Peekable;
-
 use pubgrub::range::Range;
 use tracing::debug;
 
@@ -13,6 +9,7 @@ use uv_configuration::IndexStrategy;
 use uv_normalize::PackageName;
 use uv_types::InstalledPackagesProvider;
 
+use crate::iterators::{MaxIterator, MinIterator};
 use crate::preferences::Preferences;
 use crate::prerelease_mode::PreReleaseStrategy;
 use crate::resolution_mode::ResolutionStrategy;
@@ -475,102 +472,5 @@ impl Name for Candidate<'_> {
 impl DistributionMetadata for Candidate<'_> {
     fn version_or_url(&self) -> distribution_types::VersionOrUrl {
         distribution_types::VersionOrUrl::Version(self.version)
-    }
-}
-
-/// An iterator that returns the maximum version from a set of iterators.
-struct MaxIterator<'a, T: Iterator<Item = (&'a Version, VersionMapDistHandle<'a>)>> {
-    iterators: Vec<Peekable<T>>,
-    heap: BinaryHeap<(&'a Version, Reverse<usize>)>,
-}
-
-impl<'a, T: Iterator<Item = (&'a Version, VersionMapDistHandle<'a>)>> MaxIterator<'a, T> {
-    fn new(iterators: Vec<T>) -> Self {
-        // Convert each iterator into a peekable.
-        let mut iterators = iterators
-            .into_iter()
-            .map(Iterator::peekable)
-            .collect::<Vec<_>>();
-
-        // Create a binary heap to track the maximum version from each iterator, respecting the
-        // order of the indexes.
-        let heap = iterators
-            .iter_mut()
-            .enumerate()
-            .filter_map(|(index, iter)| iter.peek().map(|(version, _)| (*version, Reverse(index))))
-            .collect();
-
-        Self { iterators, heap }
-    }
-}
-
-impl<'a, T: Iterator<Item = (&'a Version, VersionMapDistHandle<'a>)>> Iterator
-    for MaxIterator<'a, T>
-{
-    type Item = (&'a Version, VersionMapDistHandle<'a>);
-
-    fn next(&mut self) -> Option<Self::Item> {
-        // Take the maximum version from the heap.
-        let (_, index) = self.heap.pop()?;
-
-        // Advance the iterator and push the next version onto the heap.
-        let next = self.iterators[index.0].next()?;
-
-        // Push the next version onto the heap.
-        if let Some((version, _)) = self.iterators[index.0].peek() {
-            self.heap.push((version, index));
-        }
-
-        Some(next)
-    }
-}
-
-/// An iterator that returns the minimum version from a set of iterators.
-struct MinIterator<'a, T: Iterator<Item = (&'a Version, VersionMapDistHandle<'a>)>> {
-    iterators: Vec<Peekable<T>>,
-    heap: BinaryHeap<(Reverse<&'a Version>, Reverse<usize>)>,
-}
-
-impl<'a, T: Iterator<Item = (&'a Version, VersionMapDistHandle<'a>)>> MinIterator<'a, T> {
-    fn new(iterators: Vec<T>) -> Self {
-        // Convert each iterator into a peekable.
-        let mut iterators = iterators
-            .into_iter()
-            .map(Iterator::peekable)
-            .collect::<Vec<_>>();
-
-        // Create a binary heap to track the minimum version from each iterator, respecting the
-        // order of the indexes.
-        let heap = iterators
-            .iter_mut()
-            .enumerate()
-            .filter_map(|(index, iter)| {
-                iter.peek()
-                    .map(|(version, _)| (Reverse(*version), Reverse(index)))
-            })
-            .collect();
-
-        Self { iterators, heap }
-    }
-}
-
-impl<'a, T: Iterator<Item = (&'a Version, VersionMapDistHandle<'a>)>> Iterator
-    for MinIterator<'a, T>
-{
-    type Item = (&'a Version, VersionMapDistHandle<'a>);
-
-    fn next(&mut self) -> Option<Self::Item> {
-        // Take the minimum version from the heap, skipping duplicates.
-        let (_, index) = self.heap.pop()?;
-
-        // Advance the iterator and push the next version onto the heap.
-        let next = self.iterators[index.0].next()?;
-
-        // Push the next version onto the heap.
-        if let Some((version, _)) = self.iterators[index.0].peek() {
-            self.heap.push((Reverse(version), index));
-        }
-
-        Some(next)
     }
 }
